create d_a_t_wood_area header
